requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.1.3
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.1
pyee==9.0.4
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
slackclient==2.9.4
slackeventsapi==3.0.1
tqdm==4.65.0
typing-extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.15
Werkzeug==2.2.0
yarl==1.7.2
zipp==3.8.1
